+++ src/index.ts
@@ -30,6 +30,7 @@ export * from '@project-chip/matter-node.js/log';
 export * from '@project-chip/matter-node.js/datatype';
 export * from '@project-chip/matter.js/util';
 export * from '@project-chip/matter.js/schema';
+export * from '@project-chip/matter.js/tlv';
 
 export * from 'matter-history';